Add subtle syntax highlighting when in json editing mode on endpoint body editor
The tokeniser is very lenient, it parses line and block comments rather than rejecting them as some JSON implementations may, and it allows multiple json values in the body rather than requiring exactly one, and may allow malformed numbers. The goal of this subtle highlighting is to make it easier to identify where a syntax error was introduced, by making it clearer where strings for keys or values end. We might want to add different colours in the future, but for now it just uses different alpha levels to provide emphasis.